pyproject.toml
[project]
name = "statsig-openfeature-provider-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openfeature-sdk>=0.8.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/statsig_openfeature_provider_python"]

[dependency-groups]
dev = [
    "pytest",
]
